requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==35.0.0
Django==4.1.5
django-appconf==1.0.5
django-easy-tenants==0.7.1
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
Markdown==3.4.1
multidict==6.0.4
mypy-extensions==0.4.3
mysqlclient==2.1.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
pathspec==0.10.3
pick==2.2.0
platformdirs==2.6.2
progressbar==2.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
versioneer==0.28
yarl==1.8.2
